+++ src/app/login/page.tsx
@@ -4,16 +4,12 @@ import LoginForm from '@/components/login/LoginForm';
 
 export default async function LoginPage() {
   const supabase = createClient();
-  const { data, error } = await supabase.auth.getUser();
-  const session = await supabase.auth.getSession();
+  const { data } = await supabase.auth.getUser();
 
-  console.log('session => ', session);
-  console.log('data => ', data);
-
-  // if (data.user) {
-  //   redirect('/myclub');
-  //   return null;
-  // }
+  if (data.user) {
+    redirect('/myclub');
+    return null;
+  }
   return (
     <div className='bg-mainblue h-full'>
       <LoginForm />
